Выделил масштаб и эффекты в отдельные модули
