+++ js/main.js
@@ -1,3 +1,5 @@
 import './gallery.js';
 import './big-picture.js';
+import './scale.js';
+import './effects.js';
 import './upload-picture.js';
